+++ src/trsp/trsp_driver.cpp
@@ -202,7 +202,7 @@ pgr_do_trsp(
         auto new_combinations = pgrouting::utilities::get_combinations(paths, ruleList);
 
         if (!new_combinations.empty()) {
-            pgrouting::trsp::Pgr_trspHandler gdef(
+            pgrouting::trsp::TrspHandler gdef(
                     edges,
                     directed,
                     ruleList);